Update README.md
This repository presents an analysis of the impact of a CO2 emissions cap on the energy system configuration. The assignment at hand involves illustrating three scenarios (low, medium, high) and comparing the resulting changes in the energy system.

The repository contains Jupyter notebooks that provide an exploration of the problem, including discussions on the intuition behind the results. The notebooks are designed to go beyond a single solution, offering insights into 3 different scenarios and their respective outcomes.

To ensure seamless execution, the repository includes clear instructions on how to run the notebooks in the required order. Additionally, it accommodates the local execution of MESSAGEix, a powerful modeling framework, by utilizing a GAMS license.

By accessing this repository, you will gain insights into the effects of implementing a CO2 emissions cap and its implications for the energy system. The notebook code needs a valid GAMS license to fully leverage the capabilities of MESSAGEix and reproduce the results effectively.

**Every step towards analyzing and mitigating CO2 emissions brings us closer to a more sustainable and resilient world. Happy exploring and analyzing!**

Note: This repository requires a valid GAMS license for local execution of MESSAGEix.